+++ instrumentation/github.com/99designs/gqlgen/otelgqlgen/config.go
@@ -1,4 +1,18 @@
-package graphql
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package otelgqlgen
 
 import (
 	"go.opentelemetry.io/otel/propagation"
@@ -6,8 +20,8 @@ import (
 )
 
 type config struct {
-	TracerProvider trace.TracerProvider
-	Propagators    propagation.TextMapPropagator
+	tracerProvider trace.TracerProvider
+	propagators    propagation.TextMapPropagator
 }
 
 // Option specifies instrumentation configuration options.
@@ -26,7 +40,7 @@ func (o optionFunc) apply(c *config) {
 // ones will be used.
 func WithPropagators(propagators propagation.TextMapPropagator) Option {
 	return optionFunc(func(cfg *config) {
-		cfg.Propagators = propagators
+		cfg.propagators = propagators
 	})
 }
 
@@ -34,6 +48,6 @@ func WithPropagators(propagators propagation.TextMapPropagator) Option {
 // If none is specified, the global provider is used.
 func WithTracerProvider(provider trace.TracerProvider) Option {
 	return optionFunc(func(cfg *config) {
-		cfg.TracerProvider = provider
+		cfg.tracerProvider = provider
 	})
 }