+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - Add the `zpages` span processor. (#894)
+- `99designs/gqlgen` trace instrumentation. (#951)
 
 ### Changed